Update MouseEvents for grid size and size label
Mouseevent 'change' is being used instead of 'input' to solve lag
that happens when sliding the slide bar. Before, the grid size will
consistently update when the bar is being dragged. Now, instead, it only
updates when the bar has been changed.

Separate function was created for slider label, to where it changes
when the slider is being dragged.